implimented server priority / and non-blocking scheduling. no longer will crafty "sleep" while waiting for a server start delay. we use the scheduling system to do this now. so, we create a list of high/med/low priority servers (per the db) and order them by delay so high priority, low delay servers are the absolute first, then next delay - then we do the same for medium, and then low.
when crafty launches, it will init these servers (via multi) and once the server objects are created/stored in multi - if the server auto-starts, we schedule the new "run_scheduled_server" function (in minecraft_server.py)
this task, once executed (via schedule) will execute run threaded server, and then delete it's schedule (this is how it runs only once).

this prevents the blocking behavior of crafty on launch for servers who might have a huge delay (like 120 seconds) this also prevents the dashboard errors we have been seeing with 3.0
where a server instance / object isn't created due to the delay of another server.