base.html: change behaviour while clicking on logo
While you are logged as Admin or mod, you don't expect that while
clicking on the logo of Crafty Controller, it logs out you from the UI
and throws you to Crafty Controller Login where you can see server status
and login form. Let's change this behavior to move you to the dashboard.

If you want to be logged out, just click on your Username (user-menu)
and select it there.